Moved business logic to service classes

+++ src/main/java/com/shopify/dto/InventoryDTO.java
@@ -7,5 +7,5 @@ public interface InventoryDTO {
 
 	int getQuantity();
 
-	int getPrice();
+	double getPrice();
 }
